+++ src/target/lua/misc/LuaKeyword.ts
@@ -2,8 +2,10 @@ import { Target } from "../../Target";
 import * as ts from "typescript";
 import { BaseTarget } from "../../BaseTarget";
 import { LuaKeywords } from "../LuaKeywords";
+import { LuaClassDeclaration } from "../declarations";
+import { UnsupportedError } from "../../../error/UnsupportedError";
 
-export interface LuaKeyword extends BaseTarget, Target { }
+export interface LuaKeyword extends BaseTarget, Target, LuaClassDeclaration { }
 export class LuaKeyword implements Partial<Target> {
 
     public transpileKeyword(node: ts.Node): string {
@@ -19,7 +21,13 @@ export class LuaKeyword implements Partial<Target> {
             case ts.SyntaxKind.ThisKeyword:
                 return "self";
             case ts.SyntaxKind.SuperKeyword:
-                return `self.${LuaKeywords.CLASS_SUPER_REFERENCE_NAME}`;
+
+                const className = this.lastParsedClassName;
+                if (!className) {
+                    throw new UnsupportedError(`Trying to access a super object while not beeing in a class context is unsupported!`, node);
+                }
+
+                return `${className}.${LuaKeywords.CLASS_SUPER_REFERENCE_NAME}`;
         }
     }
 }